Factorize code

+++ src/events/dyn_vocal/on_disconnect.ts
@@ -9,13 +9,10 @@ import type { BotEvent } from '@events';
 export const DYN_VOCAL_ON_DISCONNECT: BotEvent = {
     name: 'Dynamic vocal channel - on disconnect',
     listenTo: 'voiceStateUpdate',
-    async execute(oldStateChannel, _) {
-        if (oldStateChannel.channelId === null || oldStateChannel.guild == null) return;
-        const channel = oldStateChannel.channel;
-        if (!channel || !channel.name.startsWith('Nouveau')) return;
-        if (channel.members.size === 0) {
-            await channel.delete();
-        }
+    async execute({ channel }, _) {
+        if (!channel || !channel.name.startsWith('Nouveau') || channel.members.size) return;
+
+        await channel.delete();
     },
     once: false,
 };
